+++ src/main/java/net/rasanovum/rockandstone/RockAndStoneConfig.java
@@ -6,5 +6,6 @@ public class RockAndStoneConfig extends MidnightConfig {
     @Entry public static boolean doRockAndStoneDebug = false;
     @Entry public static boolean doOreReplacement = true;
     @Entry public static boolean doCreateRnsTerrainFiltering = false;
+    @Entry public static boolean doLargeOreDepositsTerrainFiltering = false;
     @Entry public static boolean doSilentAdvancements = false;
 }